Added return in case of no password input

+++ src/routes/EditUser.jsx
@@ -45,6 +45,9 @@ function EditUser() {
   const DeleteProfile = async() => {
     if (confirm("Are you sure you want to delete your account?")) {
       const password = prompt("Please enter your password to confirm deletion")
+      if (password === null) {
+        return;
+      }
       await axios.delete(
         "/user",
         {
